# backend/internal/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "EpicPanel API",
    "version": "1.0.0",
    "description": "Linux-first hosting control plane API. Session cookie or Bearer session token for users; Bearer epk_ tokens for automation (scoped)."
  },
  "servers": [{"url": "/"}],
  "components": {
    "securitySchemes": {
      "bearer": {"type": "http", "scheme": "bearer"}
    },
...
  },
  "paths": {
...
    "/v1/organizations/{org_id}/websites/{website_id}/deployment-config": {"patch": {"summary": "Set git deploy config (developer+)", "security": [{"bearer": []}], "responses": {"204": {"description": "done"}}}},
    "/v1/organizations/{org_id}/websites/{website_id}/deploy": {"post": {"summary": "Trigger deploy (developer+) -> 202", "security": [{"bearer": []}], "responses": {"202": {"description": "deploying"}}}},
    "/v1/organizations/{org_id}/websites/{website_id}/rollback": {"post": {"summary": "Rollback to last successful release (admin+) -> 202", "security": [{"bearer": []}], "responses": {"202": {"description": "rolling back"}}}},
    "/v1/organizations/{org_id}/websites/{website_id}/deployments": {"get": {"summary": "Deployment history (billing+)", "security": [{"bearer": []}], "responses": {"200": {"description": "ok"}}}},
...
  }
}